Move git auth tests to a new file
It requires a lot of helper functions and was cluttering up the
`source_test.go` file.

Also, remove some dead code and do some cleanup.
